Started basic implementation of TaCo 2 and moved old files into v1 directory
