+++ v1/diff_2d_histogram.js
