+++ v1/diff_barplot.js
