+++ v1/diff_heatmap.js
