+++ v1/diff_histogram.js
