+++ v1/diff_percentagebar.js
